No more impersonating others -- you must verify your in-game identity
Well, sort of.  By default the relay will use the discord username.
If people want to use their own names, they can send a direct
message to the bot, in KoL, asking them to 'Verify' them.

The bot will then give them a passphrase, that they must recite
to the bot in discord.  That will link both accounts.

Implementation-wise, it's a barebones sqlite3 table and two
sync hashes.  Turns out go-sqlite allows concurrent reads (yay!)
but not concurrent writes (nay!), so we do the mutexing for
the latter manually.