Enforce the 200-character limit per message
hekiryuu++ for reporting this & explaining what the limitd are, and
what other applications (e.g. mafia) do to handle it.

We take a bit of a middle ground.  If the message is over 300
characters, the bot gets sassy in discord and does not send it.
Between 200 and 300, it splits it into two.  Below 200, it just
sends it as usual.