Send a one-off message to people with nicknames
The relay uses usernames by default, to prevent abuse.

It is possible to verify your in-game name and make it use that,
but usually this requires someone to explain the process.

So now the relay will spot people with nicknames, and send them
a one-off message letting them know what's up.