Ensure the guild roles are loaded
Wow.  I don't even know how this happens, but in my local confuser,
s.Guild(id).Roles returns a list of roles, while in a remote system,
it is generally empty.  Sounds like a concurrency issue, and who knows
where (discordgo?).

This wallpapers over the issue by checking if the roles we got are
empty, and if so, going to discord for the full list.