Use the status API to check for jawbruiser & friends
This means we do 1+N requests, where N is the number of
effects that we have and must(jawbruiser, unmotivated, etc);
previously we had to do a constant K requests, where K is
all the effects we had to remove.

So in practice, we've gone from always doing 3 requests, to
always doing 1 request, and potentially having to do 3 more;
worst case scenario is slightly worse, best case scenario is
much better.

Once kolgo starts maintaining the statuses internally, we'll
be able to do 0+N requests.