Prevent a bunch of nuls from showing up in messages due to mis-allocations
Was doing this:

    f := make([]byte, 300)
    f  = append(f, 'f')

This left f with 300 nuls followed by an f.  What we really wanted
was this:

    f := make([]byte, 0, 300)
    f  = append(f, 'f')

Which does exactly what we wanted.  So this patch does exactly that.

This was reported by EstroJenn -- for some reason, discord was
displaying the nuls as [] for her, making some messages have a
gargantuan prefix of [][][]s.